Make NSPersistentStoreCoordinator methods private
