--- PeakCoreData/Helpers/NSPersistentStoreCoordinator.swift
@@ -1,42 +1,0 @@
-//
-//  NSPersistentStoreCoordinator.swift
-//  PeakCoreData
-//
-//  Created by David Yates on 10/05/2019.
-//  Copyright © 2019 3Squared Ltd. All rights reserved.
-//
-
-import CoreData
-
-extension NSPersistentStoreCoordinator {
-    
-    static func destroyStore(at storeURL: URL) {
-        do {
-            let psc = NSPersistentStoreCoordinator(managedObjectModel: NSManagedObjectModel())
-            try psc.destroyPersistentStore(at: storeURL, ofType: NSSQLiteStoreType, options: nil)
-        } catch let error {
-            fatalError("Failed to destroy persistent store at \(storeURL), error: \(error)")
-        }
-    }
-    
-    static func replaceStore(at targetURL: URL, withStoreAt sourceURL: URL) {
-        do {
-            let psc = NSPersistentStoreCoordinator(managedObjectModel: NSManagedObjectModel())
-            try psc.replacePersistentStore(at: targetURL, destinationOptions: nil, withPersistentStoreFrom: sourceURL, sourceOptions: nil, ofType: NSSQLiteStoreType)
-        } catch let error {
-            fatalError("Failed to replace persistent store at \(targetURL) with \(sourceURL), error: \(error)")
-        }
-    }
-    
-    static func metadata(at storeURL: URL) -> [String : Any]?  {
-        return try? NSPersistentStoreCoordinator.metadataForPersistentStore(ofType: NSSQLiteStoreType, at: storeURL, options: nil)
-    }
-    
-    func addPersistentStore(at storeURL: URL, options: [AnyHashable : Any]) -> NSPersistentStore {
-        do {
-            return try addPersistentStore(ofType: NSSQLiteStoreType, configurationName: nil, at: storeURL, options: options)
-        } catch {
-            fatalError("Failed to add persistent store to coordinator, error: \(error)")
-        }
-    }
-}
